// vite/src/lib/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          created_at: string;
          id: string;
          language: string;
          main_image: string | null;
          timezone: string;
          updated_at: string;
          user_email: string;
          user_id: string | null;
          username: string | null;
        };
        Insert: {
          created_at?: string;
          id?: string;
          language?: string;
          main_image?: string | null;
          timezone?: string;
          updated_at?: string;
          user_email?: string;
          user_id?: string | null;
          username?: string | null;
        };
        Update: {
          created_at?: string;
          id?: string;
          language?: string;
          main_image?: string | null;
          timezone?: string;
          updated_at?: string;
          user_email?: string;
          user_id?: string | null;
          username?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      add_daily_workouts: {
        Args: {
          body: Json;
        };
        Returns: undefined;
      };
      add_meals: {
        Args: {
          body: Json;
        };
        Returns: undefined;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};